cnf ran: remove cat hack from powersave
The powersave test cases would pipe the output from cat into `cat -` which causes errors to be ignored as a workaround for ocassional malformed output. This has been converted to us Eventually to allow for retrying on malformed output.

This PR also updates the new CPU frequency tests in a similar manner to ensure they are resillient against malformed outputs.