cnf ran: add reporting for CRs on each cluster
Currently, the cnf/ran test suites do not log the CRs used in each test case, which becomes particularly important for the TALM and ZTP test suites. This PR uses the existing infrastructure in the repo around reporting to generate dumps with these resources.

There is an additional function in the internal/reporter package that allows reports to dump the CRs from a specific kubeconfig, not just the one specified by the KUBECONFIG environment variable.