cnf ran: ztp generator tweaks based on testing (#104)
Testing the ZTP generator tests in the actual environment revealed that the output from brew was not handled properly. This PR adds a sed command to retrieve the version tag from the output.

Additionally, the no-container label is added for tests that should not be executed in a container, such as ztp-generator currently.