global: fail fast on retry errors
As originally used by the cnf/ran directory, retry is only supposed to retry on errors that come from the command execution failing, often from the MCO daemon restarting mid-execution. The current implementation will ignore errors completely that are not to be retried. This masks any errors in the actual command failing.

This PR just changes back to the original behavior and will fail fast when the command gets executed but fails, for example, by having a nonzero exit code.