rhwa: DAST test on FAR operator
Adds a test that runs rapidast scan for the FAR operator installed in the cluster.

It configures the rapidast for the execution of the trivy scan, looking for
misconfigurations on the pods running on the namespace where the operator is installed.

The test fails if any High severity vulnerability is found, providing in the
console output the details of the detection.