report: add most of report program
This PR contains about half of the report program as part of splitting it up into more managable PR sizes. It includes the following files:

* `command.go`: Wrapper around executing local commands such as git and ginkgo.
* `sum.go`: Small utility for getting the SHA-256 sum of the source code.
* `template.go`: Contains types and functions for executing the report templates.
* `tree.go`: Defines the SuiteTree type and methods for creating one from a Ginkgo report.

The excluded files are necessary to actually run the program, namely the cache and main files. However, these combined are quite large and depend on all of the files from this PR.