global: update WaitForRecover to use new method (#449)
This is part 1 of a 3 part effort to replace pod.WaitForAllPodsInNamespacesHealthy.

1. (openshift-kni/eco-goinfra#928) Add new helpers to eco-goinfra that simplify waiting for pods healthy and list on every iteration.
2. (This PR) Change the cluster.WaitForRecover function to use the new eco-goinfra helpers. Also a small change in cnf/ran/talm.
3. Remove the old helpers from eco-goinfra.

Motivation:

The old WaitForRecover function would list the pods each iteration, ensuring that pods started after calling the function will be checked as well as ignoring pods that later end up deleted. A workaround for the latter of these issues was merged in #808.

Additionally, the current pod.WaitForAllPodsInNamespacesHealthy has too many parameters and is hard to test properly. The current unit tests for it do not adequately cover all scenarios. This function only gets used in one place in eco-gotests, the aforementioned cluster.WaitForRecover, where many parameters will always be the same.