ci: run lint and dry-run/tests in parallel
Similar to openshift-kni/eco-goinfra#743, this PR splits out the linting from the rest of the makefile CI. This should cut the time in half.

Also switches over to using go.mod to get the go version, that way we do not introduce another place where the version needs to be updated.