lint,cnf ran: enable ginkgo linter for cnf/ran
This PR enables the ginkgo linter from golangci-lint but restricts it to the cnf/ran subdirectory. Since this linter covers many conventions that have been mostly enforced through code review in cnf/ran it is useful to have it enabled. However, it can be fairly strict so it has been left to teams to opt-in to using it.