cnf ran: fix edge case that causes nil prBuilder (#466)
In certain tests, the prBuilder is set then used again in AfterEach without being pulled again. When the test fails calling a function that sets prBuilder, this can lead to prBuilder being set to nil with no way to clean up after the tests. This PR addresses the issue by always pulling prBuilder in the AfterEach block.