report: add most of report program (#446)
* report: add ginkgo report templates

To break up the eco-gotests reporter into managable PRs, I have started with the templates for the report as a whole and the individual trees (branches).

report_template.html is the entrypoint for viewing the reports and simply contains a list of the branches in the report along with links to the latest commit on each branch.

tree_template.html is the report for an individual branch. It contains a tree listing out all of the specs under all of the subdirectories of tests/. Since it uses the native html details and summary tags it is entirely tab-navigable and fairly accessible.

Both templates are fairly barebones with the intention being that future PRs may add more details as desired.

* report: add most of report program

This PR contains about half of the report program as part of splitting it up into more managable PR sizes. It includes the following files:

* `command.go`: Wrapper around executing local commands such as git and ginkgo.
* `sum.go`: Small utility for getting the SHA-256 sum of the source code.
* `template.go`: Contains types and functions for executing the report templates.
* `tree.go`: Defines the SuiteTree type and methods for creating one from a Ginkgo report.

The excluded files are necessary to actually run the program, namely the cache and main files. However, these combined are quite large and depend on all of the files from this PR.