report: add rest of report program (#447)
This PR contains the rest of the report program. It includes the last two files not part of the last PR:

* `cache.go`: Caches the trees generated from each branch and revision combination. Since a dry-run may take many minutes, caching allows the program to remain performant and responsive to changes in the templates without having to redo a dry-run.
* `main.go`: Handles the command line flags and actual running of the program. It handles retrieving values from the cache or cloning and dry running as necessary.

There are some small changes to the lint config and go.mod to allow the cache to be compressed. It uses a package already vendored so does not include additional dependencies. Compression is beneficial since the cache for the main branch alone may be over a megabyte in size but only ~40 KB when compressed.

Excluded from this PR is the GitHub workflow to actually generate the report along with some high-level documentation.