report: fix cache in action (#527)
In switching from cache to artifacts for the cache, I did not update the action to specify the workflow run to get the cache from. Although it will not fail the workflow, this causes it to take much longer than necessary. This PR just adds a step to get the last run using the GitHub CLI.

Additionally, this switches the report to use go.mod instead of hard coding the go version. This has the effect of bumping the go version to 1.24 for generating the report.