cnf ran: refactor gitopsztp to use new git methods
Related to openshift-kni/eco-goinfra#1039, this commit refactors the gitopsztp suite to incorporate the new methods introduced for the ApplicationBuilder. Although it is slightly longer since the checking for existence is broken apart from updating and waiting for sync, this allows some more flexibility for cases where this is necessary.

The gitdetails package is left although it has a significantly reduced scope. Since the goal is to allow better reuse of code relating to ZTP, the smaller gitdetails is ultimately beneficial.