cnf ran: fix automation issue due to time ambiguity (#616)
O-RAN post-provision test case 77375 regularly fails since it can complete within a single second and WaitForPhaseAfter produces false negatives in this situation. Since the precision of times on the cluster cannot be changed, this PR solves the issue by waiting to ensure the ProvisioningRequest will not be updated in the same second as the start time.