cnf ran,system-tests: update oran tests for Metal3
Depends-on: openshift-kni/eco-goinfra#1077

This PR makes all of the changes required to support the new Metal3 hardware plugin for O-RAN tests. To avoid breaking `go vet` across multiple commits, this single PR includes changes to both the system-tests o-cloud tests and the cnf/ran oran tests. All that has been changed is the names of the Node and NodePool resources. Functionality of these tests remains the same. However, they now must be run using the Metal3 hardware plugin and will no longer support older builds of the oran-o2ims operator.

After testing, the NewNoTemplatePR method has been updated to use an invalid hostname to prevent any collisions with the existing BMH in the inventory.

Additionally, the reporter is updated so that there are no longer any references to the old oran-hwmgr-plugin repo.