gitopsztp: fix issue with node delete path (#712)
Caused by #606.

Although that PR correctly identified that appending the add suppression path would cause an error since the crAnnotation path had already been appended, it failed to account for the existence check which would be affected by the same issue. This PR just moves the fix of that issue to before the git path is checked for existence.